traits can be focus traits
The focus trait is a more extreme version of the normal trait.
They can not be chosen when a user creates their character.
A normal trait can be chosen as focustrait, it will then be replaced with the actual focus trait, the user won't know beforehand which trait it will be, he can deduct it tho.